enabled caching inside RepositoryLocator and refactoring
