+++ backend/src/main/kotlin/outcobra/server/OutstandingCobraServerApplication.kt
@@ -2,10 +2,12 @@ package outcobra.server
 
 import org.springframework.boot.SpringApplication
 import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.cache.annotation.EnableCaching
 import springfox.documentation.swagger2.annotations.EnableSwagger2
 
 @SpringBootApplication
 @EnableSwagger2
+@EnableCaching
 open class OutstandingCobraServerApplication
 
 fun main(args: Array<String>) {
