+++ backend/src/main/kotlin/outcobra/server/util/RepositoryLocator.kt
@@ -3,9 +3,11 @@ package outcobra.server.util
 import org.springframework.beans.BeansException
 import org.springframework.beans.factory.BeanFactory
 import org.springframework.beans.factory.NoSuchBeanDefinitionException
+import org.springframework.cache.annotation.Cacheable
 import org.springframework.context.ApplicationContext
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Component
+import outcobra.server.exception.NoRepositoryFoundException
 import javax.inject.Inject
 
 /**
@@ -24,6 +26,7 @@ open class RepositoryLocator @Inject constructor(val context: ApplicationContext
      * @see [BeanFactory.getBean]
      * @since 1.0.0
      */
+    @Cacheable("repoForName")
     fun getForEntityName(entityName: String): JpaRepository<*, Long> {
         val repoName = entityName.firstToLower() + "Repository"
 
@@ -46,22 +49,10 @@ open class RepositoryLocator @Inject constructor(val context: ApplicationContext
      * @throws NoRepositoryFoundException if the requested repository could not be found or has an illegal type
      * @since 1.0.0
      */
+    @Cacheable("RepoForEntity")
     fun <T> getForEntityClass(entityClass: Class<T>): JpaRepository<T, Long> {
         @Suppress("UNCHECKED_CAST")
         return getForEntityName(entityClass.simpleName) as? JpaRepository<T, Long>
                 ?: throw NoRepositoryFoundException("Could not cast repository to JpaRepository<${entityClass.simpleName}, Long>")
     }
-
-    private fun String.firstToLower(): String {
-        if (this.isNullOrEmpty()) return this
-        return substring(0, 1).toLowerCase() + substring(1, length)
-    }
 }
-
-/**
- * Is thrown when a repository can not be found or has an invalid type
- *
- * @author Joel Messerli
- * @since 1.0.0
- */
-class NoRepositoryFoundException(message: String, cause: Throwable? = null) : Exception(message, cause)
