--- backend/src/main/kotlin/outcobra/server/util/StringUtil.kt
@@ -1,13 +1,0 @@
-package outcobra.server.util
-
-/**
- * Upper-Cases the first character of a string
- *
- * @author Joel Messerli
- * @since 1.0.0
- */
-fun String.firstToUpper(): String {
-    if (this.isEmpty()) return this
-    if (this.length == 1) return this.toUpperCase()
-    return this.substring(0, 1).toUpperCase() + this.substring(1, this.length)
-}
